COMP: Rename Module to SlicerDWD
On Windows the module fails to load since `import dwd` is a circular import (`DWD.py` is case insensitive on Windows).

This renames the module and corresponding files to `SlicerDWD` so that the circular import is broken (the module is now at `SlicerDWD.py`) and the `dwd` package can be imported correctly.

Resolves #2.